Refactored the entire project for better preview and extendability
