+++ .idea/workspace.xml
@@ -1,10 +1,33 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
   <component name="ChangeListManager">
-    <list default="true" id="ad6b76c0-59b7-4bf4-9448-d2a6b18bda1a" name="Changes" comment="Dependency resolution multiple Python versions.">
+    <list default="true" id="ad6b76c0-59b7-4bf4-9448-d2a6b18bda1a" name="Changes" comment="setup.cfg fixed for URL, Author and Author Email.">
+      <change afterPath="$PROJECT_DIR$/table2xsv/bin/main.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/.gitignore" beforeDir="false" afterPath="$PROJECT_DIR$/.gitignore" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/main.py" beforeDir="false" afterPath="$PROJECT_DIR$/main.py" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/setup.cfg" beforeDir="false" afterPath="$PROJECT_DIR$/setup.cfg" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/table2xsv/__init__.py" beforeDir="false" afterPath="$PROJECT_DIR$/table2xsv/__init__.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/table2xsv/core/__init__.py" beforeDir="false" afterPath="$PROJECT_DIR$/table2xsv/core/__init__.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/table2xsv/core/base.py" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/table2xsv/core/exceptions.py" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/table2xsv/core/formatter.py" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/table2xsv/core/management/__init__.py" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/table2xsv/core/management/helpers.py" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/table2xsv/core/management/utility.py" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/table2xsv/core/parser.py" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/table2xsv/core/wrapper.py" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/table2xsv/extended/__init__.py" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/table2xsv/extended/base.py" beforeDir="false" afterPath="$PROJECT_DIR$/table2xsv/management/base.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/table2xsv/main.py" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/table2xsv/management/__init__.py" beforeDir="false" afterPath="$PROJECT_DIR$/table2xsv/management/__init__.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/table2xsv/management/commands/csv.py" beforeDir="false" afterPath="$PROJECT_DIR$/table2xsv/management/commands/csv.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/table2xsv/management/commands/excel.py" beforeDir="false" afterPath="$PROJECT_DIR$/table2xsv/management/commands/excel.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/table2xsv/management/commands/mysql.py" beforeDir="false" afterPath="$PROJECT_DIR$/table2xsv/management/commands/mysql.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/table2xsv/management/commands/neo4j.py" beforeDir="false" afterPath="$PROJECT_DIR$/table2xsv/management/commands/neo4j.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/table2xsv/management/commands/sqlite.py" beforeDir="false" afterPath="$PROJECT_DIR$/table2xsv/management/commands/sqlite.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/table2xsv/utils/getter.py" beforeDir="false" afterPath="$PROJECT_DIR$/table2xsv/utils/getter.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/tests/tests.py" beforeDir="false" afterPath="$PROJECT_DIR$/tests/tests.py" afterDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
@@ -19,6 +42,11 @@
     </option>
   </component>
   <component name="Git.Settings">
+    <option name="RECENT_BRANCH_BY_REPOSITORY">
+      <map>
+        <entry key="$PROJECT_DIR$" value="main" />
+      </map>
+    </option>
     <option name="RECENT_GIT_ROOT_PATH" value="$PROJECT_DIR$" />
   </component>
   <component name="GitSEFilterConfiguration">
@@ -44,7 +72,11 @@
   </component>
   <component name="RecentsManager">
     <key name="MoveFile.RECENT_KEYS">
-      <recent name="C:\Users\amogh.madan\PycharmProjects\Table2XSV\tests" />
+      <recent name="C:\Users\amogh.madan\PycharmProjects\Table2XSV\table2xsv" />
+      <recent name="C:\Users\amogh.madan\PycharmProjects\Table2XSV\table2xsv\contrib" />
+      <recent name="C:\Users\amogh.madan\PycharmProjects\Table2XSV\table2xsv\contrib\root" />
+      <recent name="C:\Users\amogh.madan\PycharmProjects\Table2XSV\table2xsv\bin" />
+      <recent name="C:\Users\amogh.madan\PycharmProjects\Table2XSV\table2xsv\management" />
     </key>
   </component>
   <component name="RunManager" selected="tox.TOX (Build): Package">
@@ -168,7 +200,14 @@
       <option name="project" value="LOCAL" />
       <updated>1636522369034</updated>
     </task>
-    <option name="localTasksCounter" value="15" />
+    <task id="LOCAL-00015" summary="setup.cfg fixed for URL, Author and Author Email.">
+      <created>1636523364318</created>
+      <option name="number" value="00015" />
+      <option name="presentableId" value="LOCAL-00015" />
+      <option name="project" value="LOCAL" />
+      <updated>1636523364318</updated>
+    </task>
+    <option name="localTasksCounter" value="16" />
     <servers />
   </component>
   <component name="Vcs.Log.Tabs.Properties">
@@ -197,6 +236,7 @@
     <MESSAGE value="Fixed Doc String for handle method" />
     <MESSAGE value="Update dependencies to the latest version." />
     <MESSAGE value="Dependency resolution multiple Python versions." />
-    <option name="LAST_COMMIT_MESSAGE" value="Dependency resolution multiple Python versions." />
+    <MESSAGE value="setup.cfg fixed for URL, Author and Author Email." />
+    <option name="LAST_COMMIT_MESSAGE" value="setup.cfg fixed for URL, Author and Author Email." />
   </component>
 </project>
