+++ main.py
@@ -1,6 +1,6 @@
-from table2xsv.main import main
+from table2xsv.bin import execute_from_command_line
 
 if __name__ == "__main__":
     """Run"""
 
-    main()
+    execute_from_command_line()
