+++ setup.cfg
@@ -41,7 +41,7 @@ install_requires =
 
 [options.entry_points]
 console_scripts =
-    Table2XSV = table2xsv.main:main
+    Table2XSV = table2xsv.bin:execute_from_command_line
 
 [options.extras_require]
 all =
