+++ table2xsv/bin/__init__.py
@@ -1,0 +1,5 @@
+from table2xsv.bin.main import execute_from_command_line
+
+__all__ = [
+    "execute_from_command_line",
+]
