+++ table2xsv/core/__init__.py
@@ -1,7 +1,10 @@
 from table2xsv.core.base import BaseCommand
-from table2xsv.core.exceptions import CommandError
+from table2xsv.core.exceptions import CommandError, SystemCheckError
+from table2xsv.core.utility import ManagementUtility
 
 __all__ = [
     "BaseCommand",
     "CommandError",
+    "SystemCheckError",
+    "ManagementUtility",
 ]
