+++ table2xsv/core/helpers.py
@@ -3,7 +3,8 @@ import functools
 import pkgutil
 from importlib import import_module
 
-from table2xsv.core.base import BaseCommand, CommandError
+from table2xsv.core.base import BaseCommand
+from table2xsv.core.exceptions import CommandError
 from table2xsv.utils.getter import get_package_path
 
 _AppendConstAction = getattr(argparse, "_AppendConstAction")
