--- table2xsv/extended/__init__.py
@@ -1,5 +1,0 @@
-from table2xsv.extended.base import Table2XSVBaseCommand
-
-__all__ = [
-    "Table2XSVBaseCommand",
-]
