+++ table2xsv/management/commands/excel.py
@@ -2,7 +2,7 @@ import sys
 
 from pandas import read_excel
 
-from table2xsv.extended.base import Table2XSVBaseCommand
+from table2xsv.management import Table2XSVBaseCommand
 
 
 class Command(Table2XSVBaseCommand):
