+++ table2xsv/management/commands/neo4j.py
@@ -8,7 +8,7 @@ except ModuleNotFoundError:
     sys.exit(msg)
 from pandas import DataFrame
 
-from table2xsv.extended.base import Table2XSVBaseCommand
+from table2xsv.management import Table2XSVBaseCommand
 
 
 class Command(Table2XSVBaseCommand):
